serverstuff

+++ src/java/server/Device.java
@@ -14,8 +14,8 @@ import java.util.ArrayList;
 class Device {
     private String name;
     private final String id;
-    private float x;
-    private float y;
+    private double x;
+    private double y;
     private ArrayList<Float> history;//HISTORICAL COORDINATES
     private float direction;
     
@@ -41,19 +41,19 @@ class Device {
         return this.id;
     }
     
-    public float getX() {
+    public double getX() {
         return this.x;
     }
     
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
     
-    public float getY() {
+    public double getY() {
         return this.y;
     }
     
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
